requirements.txt
asgiref==3.2.3
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.6
django-extensions==3.1.1
django-widget-tweaks==1.4.5
djangorestframework==3.12.2
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.16
Pygments==2.8.0
pytz==2019.3
requests==2.25.1
sqlparse==0.3.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
